cleanup2

+++ Testing_Reloaded_Client/TestManager.cs
@@ -1,44 +1,37 @@
 ﻿using System;
-using System.Diagnostics;
 using System.IO;
-using System.Net.Sockets;
-using System.Runtime.InteropServices;
 using System.Threading.Tasks;
 using ICSharpCode.SharpZipLib.Zip;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
-using SharedLibrary;
 using SharedLibrary.Models;
-using SharedLibrary.Statics;
 using Testing_Reloaded_Client.Networking;
 using static SharedLibrary.Models.UserTestState;
 
 namespace Testing_Reloaded_Client {
     public class TestManager {
-        private User me;
-        private NetworkManager netManager;
-
-        private Test currentTest;
+        private readonly User me;
+        private readonly NetworkManager netManager;
         public UserTestState TestState;
-
-        public Test CurrentTest => currentTest;
-
-        public string ResolvedTestPath => this.ResolvePath(this.currentTest.ClientTestPath);
-
-        public event Action ReloadUI;
 
         public TestManager(Server server, User me) {
             this.me = me;
-            this.netManager = new NetworkManager(server);
+            netManager = new NetworkManager(server);
             netManager.ReceivedMessageFromServer += ReceivedServerMessage;
         }
 
+        public Test CurrentTest { get; private set; }
+
+        public string ResolvedTestPath => ResolvePath(CurrentTest.ClientTestPath);
+
+        public event Action ReloadUI;
+
         private string ReceivedServerMessage(Server s, JObject message) {
             if (message["Action"].ToString() == "UpdateTest") {
-                this.currentTest = (Test) message["Test"].ToObject(typeof(Test));
+                CurrentTest = (Test) message["Test"].ToObject(typeof(Test));
 
-                if (currentTest.State == Test.TestState.OnHold)
-                    this.TestState.State = UserState.OnHold;
+                if (CurrentTest.State == Test.TestState.OnHold)
+                    TestState.State = UserState.OnHold;
 
                 SendStateUpdate();
                 ReloadUI?.Invoke();
@@ -51,7 +44,7 @@ namespace Testing_Reloaded_Client {
 
         public string ResolvePath(string path) {
             return Environment.ExpandEnvironmentVariables(path).Replace("$surname", me.Surname)
-                .Replace("$test_name", currentTest.TestName);
+                .Replace("$test_name", CurrentTest.TestName);
         }
 
         public async Task Connect() {
@@ -63,35 +56,34 @@ namespace Testing_Reloaded_Client {
         public async Task DownloadTestData() {
             var packet = new {Action = "GetTestInfo"};
 
-            var str = (JsonConvert.SerializeObject(packet));
+            var str = JsonConvert.SerializeObject(packet);
 
             await netManager.WriteLine(str);
 
             var response = await netManager.ReadLine();
             var jsonResponse = JObject.Parse(response);
 
-            this.currentTest = JsonConvert.DeserializeObject<Test>(jsonResponse["Test"].ToString());
-            TestState = new UserTestState {RemainingTime = currentTest.Time, State = UserState.Waiting};
+            CurrentTest = JsonConvert.DeserializeObject<Test>(jsonResponse["Test"].ToString());
+            TestState = new UserTestState {RemainingTime = CurrentTest.Time, State = UserState.Waiting};
         }
 
         public async Task WaitForTestStart() {
-            if (currentTest.State != Test.TestState.NotStarted) return;
+            if (CurrentTest.State != Test.TestState.NotStarted) return;
 
             await SendStateUpdate();
 
             while (true) {
-                string response = await netManager.ReadLine();
+                var response = await netManager.ReadLine();
 
                 var jsonO = JObject.Parse(response);
                 if (jsonO["Action"].ToString() != "TestStarted") continue;
 
-                currentTest.State = Test.TestState.Started;
+                CurrentTest.State = Test.TestState.Started;
                 break;
             }
         }
 
-        public async Task Disconnect()
-        {
+        public async Task Disconnect() {
             await netManager.Disconnect();
         }
 
@@ -105,7 +97,7 @@ namespace Testing_Reloaded_Client {
 
             // await SendStateUpdate();
 
-            string path = ResolvePath(currentTest.ClientTestPath);
+            var path = ResolvePath(CurrentTest.ClientTestPath);
             if (Directory.Exists(path))
                 Directory.Delete(path, true);
             Directory.CreateDirectory(path);
@@ -132,14 +124,14 @@ namespace Testing_Reloaded_Client {
         }
 
         public async Task SendStateUpdate() {
-            string json = JsonConvert.SerializeObject(new {Action = "StateUpdate", State = TestState});
+            var json = JsonConvert.SerializeObject(new {Action = "StateUpdate", State = TestState});
 
             await netManager.WriteLine(json);
         }
 
         // must be called every 1 second
         public void TimeElapsed(uint seconds) {
-            if (currentTest.State == Test.TestState.Started)
+            if (CurrentTest.State == Test.TestState.Started)
                 TestState.RemainingTime -= TimeSpan.FromSeconds(seconds);
 
             if (TestState.RemainingTime.Seconds == 0)
@@ -148,7 +140,7 @@ namespace Testing_Reloaded_Client {
 
         public async Task Handover() {
             netManager.ProcessMessages = false;
-            
+
             await netManager.WriteLine(JsonConvert.SerializeObject(new {Action = "TestHandover"}));
 
             var fastZip = new FastZip();
@@ -161,18 +153,13 @@ namespace Testing_Reloaded_Client {
 
             var json = JObject.Parse(await netManager.ReadLine());
 
-            if (json["Status"].ToString() != "OK") {
-                throw new Exception("Server error");
-            }
+            if (json["Status"].ToString() != "OK") throw new Exception("Server error");
 
-            if (CurrentTest.DeleteFilesAfterEnd) {
-                Directory.Delete(ResolvedTestPath, true);
-            }
-
+            if (CurrentTest.DeleteFilesAfterEnd) Directory.Delete(ResolvedTestPath, true);
         }
 
         public void TestRunning() {
-           // netManager.ProcessMessages = true;
+            // netManager.ProcessMessages = true;
         }
     }
 }
